requirements.txt
Django==3.1.5
djangorestframework==3.12.1
drf-nested-routers==0.92.5
pytz==2020.1
uwsgi==2.0.19.1
pykafka==2.0.0
